Move console-commander into /shared/utility
