+++ src/core/shared/utility/console-commander.ts
